change some skill msg and fix some role errors

+++ parsian_util/include/parsian_util/action/autogenerate/kickaction.h
@@ -23,7 +23,7 @@ public:
     SkillProperty(KickAction, quint8, Robot_Id, robot_id);
     SkillProperty(KickAction, double, Tolerance, tolerance);
     SkillProperty(KickAction, bool, Chip, chip);
-    SkillProperty(KickAction, qint32, Kickspeed, kickSpeed);
+    SkillProperty(KickAction, double, Kickspeed, kickSpeed);
     SkillProperty(KickAction, qint32, Spin, spin);
     SkillProperty(KickAction, bool, Slow, slow);
     SkillProperty(KickAction, bool, Avoidpenaltyarea, avoidPenaltyArea);
