+++ parsian_util/include/parsian_util/action/autogenerate/onetouchaction.h
@@ -22,7 +22,7 @@ public:
 
     SkillProperty(OnetouchAction, quint8, Robot_Id, robot_id);
     SkillProperty(OnetouchAction, bool, Chip, chip);
-    SkillProperty(OnetouchAction, qint32, Kickspeed, kickSpeed);
+    SkillProperty(OnetouchAction, double, Kickspeed, kickSpeed);
     SkillProperty(OnetouchAction, bool, Shottoemptyspot, shotToEmptySpot);
     SkillProperty(OnetouchAction, Vector2D, Waitpos, waitPos);
     SkillProperty(OnetouchAction, Vector2D, Target, target);
